[Feat](nereids) nereids add alter view
